Break out sops decrypt script
Added a separate script file for decrypting sets of sops encrypted files. This can now
be used independently of shipping those files directly to kubernetes with k8s-deploy-secrets.
All you need to do is provide the script with a list of files and it will return a list of
decrypted file names. You can then choose to do what you'd like with that list.